+++ HISTORY.rst
@@ -1,10 +1,16 @@
 History
 =======
 
+v2.x.x (2015-04-14)
+~~~~~~~~~~~~~~~~~~~
+* Stats: "proyectos con dictamen pero sin votación"
+
 v2.0.1 (2015-03-31)
 ~~~~~~~~~~~~~~~~~~~
 * Fixed with of stats counter.
 * Better search engine. Now query words will match accented and unaccented words.
+* Stats: "en comisión sin dictamen"
+* Stats: "sin tramitar"
 
 v2.0.0 (2015-03-30)
 ~~~~~~~~~~~~~~~~~~~